requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
